review: avoid un-needed sticky module
