+++ crates/ruma-client-api/src/message/send_message_event.rs
@@ -14,7 +14,7 @@ pub mod v3 {
         MilliSecondsSinceUnixEpoch, OwnedEventId, OwnedRoomId, OwnedTransactionId,
     };
     #[cfg(feature = "unstable-msc4354")]
-    use ruma_events::sticky::StickyDurationMs;
+    use ruma_events::StickyDurationMs;
     use ruma_events::{AnyMessageLikeEventContent, MessageLikeEventContent, MessageLikeEventType};
     use serde_json::value::to_raw_value as to_raw_json_value;
